Take messages dispatching thread out from ReplicaImp - WIP (#333)
* Take messages dispatching thread out from ReplicaImp

* Fix timers_test; change replica stop order.

* Change start objects order

* Wait until dispatching thread starts

* Fix review comments

* Add test for repetitive replica stop/start. Replace Assert by exception to enable catch.

* Show replicaId correctly in the logs

+++ bftengine/src/bftengine/MsgReceiver.cpp
@@ -27,7 +27,7 @@ void MsgReceiver::onNewMessage(const NodeNum sourceNode, const char *const messa
   auto *msgBody = (MessageBase::Header *)std::malloc(messageLength);
   memcpy(msgBody, message, messageLength);
 
-  auto node = (uint16_t)sourceNode;  // TODO(GG): make sure that this casting is okay
+  auto node = sourceNode;
 
   std::unique_ptr<MessageBase> pMsg(new MessageBase(node, msgBody, messageLength, true));
